Have Vim try to colors best for a dark background.
Since I load a colorscheme within .gvimrc and not .vimrc, 'background'
is no longer set to "dark" when running in a terminal. I practically
always use a black background for my terminal windows, so this should
make Vim look a little better when I'm not running in a GUI.

Hat tip to Steven Pritchard <https://github.com/silug> for reminding me
to finally add this setting.