Keep the sign column open for certain filetypes.
This is kind of an experiment right now, to go along with my Syntastic
configuration. Syntastic makes heavy use of the sign column, but by
default, Vim only shows the sign column when there's a sign defined and
placed for the current buffer. This means the text in the buffer shifts
back and forth as the sign column is shown and hidden, and that's really
distracting. This is an attempt at fixing that. It might become a plugin
(or, more likely, I'll find a plugin that already does this and replace
my homegrown hack with it).